[dbnode] Add accessors for endpoint value in context
