+++ src/dbnode/network/server/tchannelthrift/types.go
@@ -18,6 +18,8 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
 
+// Package tchannelthrift contains code for servicing RPC requests via
+// tchannel.
 package tchannelthrift
 
 import (
@@ -56,28 +58,6 @@ const (
 	// Query represents the Query endpoint.
 	Query
 )
-
-// String returns the string value of Endpoint enum.
-func (e Endpoint) String() string {
-	switch e {
-	case AggregateRaw:
-		return "AggregateRaw"
-	case Fetch:
-		return "Fetch"
-	case FetchBatchRaw:
-		return "FetchBatchRaw"
-	case FetchBatchRawV2:
-		return "FetchBatchRawV2"
-	case FetchTagged:
-		return "FetchTagged"
-	case Query:
-		return "Query"
-	case Unknown:
-		fallthrough
-	default:
-		return "Unknown"
-	}
-}
 
 // Options controls server behavior
 type Options interface {
